Implement the encoder part of the AE model using the convolution formula - reduce the dimensions to size z.shape = (B, 32, 1, 1), after flattening z.shape = (B, 32).
Implement a loss function of your choice.

Implement the decoder part of the AE model using the transposed convolution formula. Also remember the range in which the result's value should be.

Implement a DAE. In the dataset section of the program, use torch.random.random(shape) or another function of your choice to ensure that each pixel x in the input data is replaced with a value of 0 in 50% of the cases. Use the original uncorrupted image for 50% of the input samples as well.